experimento_disponibilidad/recovery-payment/requirements.txt
Flask==3.0.2
flask-marshmallow==1.2.0
Flask-RESTful==0.3.10
requests==2.31.0
pika==1.3.2
requests==2.31.0
rq==1.15.1
stomp.py==8.1.0